+++ src/contailners/doctors/doctor.js
@@ -6,7 +6,7 @@ import {
   Loader, Image, Card,
 } from 'semantic-ui-react';
 import { fetchDoctorShowStartAsync } from '../../actions/doctors/doctorShowActions';
-import DoctorAppointment from '../../components/doctorAppointment/doctorAppointment';
+import DoctorAppointment from '../../components/doctors/doctorAppointment';
 
 const Doctor = (props) => {
   const doctor = useSelector((state) => state.doctorShow);
